requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
dj-database-url==0.5.0
dj-pagination==2.4.0
Django==2.2
django-admin-rangefilter==0.5.0
django-autoslug==1.9.6
django-filter==2.2.0
django-money==0.15
django-multiselectfield==0.1.8
django-searchable-encrypted-fields==0.1.3
django-summernote==0.8.11.4
django-widget-tweaks==1.4.5
gunicorn==19.9.0
importlib-metadata==0.22
more-itertools==7.2.0
packaging==19.1
Pillow==6.1.0
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
py-moneyed==0.8.0
pycryptodome==3.9.0
pyparsing==2.4.2
pytest==5.1.2
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
whitenoise==4.1.3
zipp==0.6.0
